fix(sql): refresh plan cache accounting CAS value
Concurrent prepared statement cleanup can lose the first managed-memory CAS. ATOMIC_BCAS does not update its expected value, so retrying with the stale value spins forever. Reload the counter before every retry and cover both plan cache paths with concurrent regression tests.

+++ src/storage/tx/ob_id_service.cpp
@@ -312,7 +312,7 @@ int ObIDService::prepare_next_number(const int64_t base_id)
       return OB_SUCCESS;
     }
     const int64_t max_pre_allocated_id = max_pre_allocated_id_(base_id);
-    const int64_t pre_allocated_step = max(10L, pre_allocated_range_);
+    const int64_t pre_allocated_step = max(static_cast<int64_t>(10), pre_allocated_range_);
     const int64_t pre_allocated_id = max_pre_allocated_id <= next_id
         || pre_allocated_step >= max_pre_allocated_id - next_id
         ? max_pre_allocated_id
